Deleted old aspnet5 sample - created a new one for RC1
