+++ source/AspNet5/AspNet5.sln
@@ -1,0 +1,46 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.23107.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{49A08B90-4127-4614-B228-CB83531CDA53}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{66F9C8BB-632E-4201-A7DB-673A57D21070}"
+	ProjectSection(SolutionItems) = preProject
+		global.json = global.json
+	EndProjectSection
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "IdentityServer", "src\IdentityServer\IdentityServer.xproj", "{219F7ED9-2215-43FC-B339-F7D2A50CE239}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MvcClient", "src\MvcClient\MvcClient.xproj", "{A57F3C0A-A8AB-411E-861F-37F620AB0FB8}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Api", "src\Api\Api.xproj", "{728F86D8-D23F-4DDD-96DE-0F9FE08BA825}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{219F7ED9-2215-43FC-B339-F7D2A50CE239}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{219F7ED9-2215-43FC-B339-F7D2A50CE239}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{219F7ED9-2215-43FC-B339-F7D2A50CE239}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{219F7ED9-2215-43FC-B339-F7D2A50CE239}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A57F3C0A-A8AB-411E-861F-37F620AB0FB8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A57F3C0A-A8AB-411E-861F-37F620AB0FB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A57F3C0A-A8AB-411E-861F-37F620AB0FB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A57F3C0A-A8AB-411E-861F-37F620AB0FB8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{728F86D8-D23F-4DDD-96DE-0F9FE08BA825}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{728F86D8-D23F-4DDD-96DE-0F9FE08BA825}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{728F86D8-D23F-4DDD-96DE-0F9FE08BA825}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{728F86D8-D23F-4DDD-96DE-0F9FE08BA825}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{219F7ED9-2215-43FC-B339-F7D2A50CE239} = {49A08B90-4127-4614-B228-CB83531CDA53}
+		{A57F3C0A-A8AB-411E-861F-37F620AB0FB8} = {49A08B90-4127-4614-B228-CB83531CDA53}
+		{728F86D8-D23F-4DDD-96DE-0F9FE08BA825} = {49A08B90-4127-4614-B228-CB83531CDA53}
+	EndGlobalSection
+EndGlobal
