+++ source/AspNet5/src/Api/Controllers/IdentityController.cs
@@ -1,0 +1,18 @@
+﻿using System.Linq;
+using Microsoft.AspNet.Mvc;
+using Microsoft.AspNet.Authorization;
+
+namespace Api.Controllers
+{
+    [Route("identity")]
+    [Authorize]
+    public class IdentityController : Controller
+    {
+        [HttpGet]
+        public IActionResult Get()
+        {
+            return Json(from c in User.Claims
+                        select new { c.Type, c.Value });
+        }
+    }
+}
