+++ source/AspNet5/src/IdentityServer/Configuration/Scopes.cs
@@ -1,0 +1,33 @@
+﻿using System.Collections.Generic;
+using IdentityServer3.Core.Models;
+
+namespace IdentityServer.Configuration
+{
+    public class Scopes
+    {
+        public static IEnumerable<Scope> Get()
+        {
+            return new[]
+            {
+                // standard OpenID Connect scopes
+                StandardScopes.OpenId,
+                StandardScopes.ProfileAlwaysInclude,
+                StandardScopes.EmailAlwaysInclude,
+
+                // API - access token will 
+                // contain roles of user
+                new Scope
+                {
+                    Name = "api1",
+                    DisplayName = "API 1",
+                    Type = ScopeType.Resource,
+
+                    Claims = new List<ScopeClaim>
+                    {
+                        new ScopeClaim("role")
+                    }
+                }
+            };
+        }
+    }
+}
