+++ source/AspNet5/src/IdentityServer/Configuration/Users.cs
@@ -1,0 +1,43 @@
+﻿using System.Collections.Generic;
+using System.Security.Claims;
+using IdentityServer3.Core;
+using IdentityServer3.Core.Services.InMemory;
+
+namespace IdentityServer.Configuration
+{
+    static class Users
+    {
+        public static List<InMemoryUser> Get()
+        {
+            var users = new List<InMemoryUser>
+            {
+                new InMemoryUser{Subject = "818727", Username = "alice", Password = "alice",
+                    Claims = new Claim[]
+                    {
+                        new Claim(Constants.ClaimTypes.Name, "Alice Smith"),
+                        new Claim(Constants.ClaimTypes.GivenName, "Alice"),
+                        new Claim(Constants.ClaimTypes.FamilyName, "Smith"),
+                        new Claim(Constants.ClaimTypes.Email, "AliceSmith@email.com"),
+                        new Claim(Constants.ClaimTypes.EmailVerified, "true", ClaimValueTypes.Boolean),
+                        new Claim(Constants.ClaimTypes.Role, "Admin"),
+                        new Claim(Constants.ClaimTypes.Role, "Geek"),
+                    }
+                },
+                new InMemoryUser{Subject = "88421113", Username = "bob", Password = "bob",
+                    Claims = new Claim[]
+                    {
+                        new Claim(Constants.ClaimTypes.Name, "Bob Smith"),
+                        new Claim(Constants.ClaimTypes.GivenName, "Bob"),
+                        new Claim(Constants.ClaimTypes.FamilyName, "Smith"),
+                        new Claim(Constants.ClaimTypes.Email, "BobSmith@email.com"),
+                        new Claim(Constants.ClaimTypes.EmailVerified, "true", ClaimValueTypes.Boolean),
+                        new Claim(Constants.ClaimTypes.Role, "Developer"),
+                        new Claim(Constants.ClaimTypes.Role, "Geek"),
+                    }
+                }
+            };
+
+            return users;
+        }
+    }
+}
