+++ source/AspNet5/src/MvcClient/Startup.cs
@@ -1,0 +1,50 @@
+﻿using Microsoft.AspNet.Builder;
+using Microsoft.AspNet.Hosting;
+using Microsoft.Extensions.DependencyInjection;
+using System.IdentityModel.Tokens.Jwt;
+
+namespace MvcClient
+{
+    public class Startup
+    {
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddMvc();
+        }
+
+        public void Configure(IApplicationBuilder app)
+        {
+            JwtSecurityTokenHandler.DefaultInboundClaimTypeMap.Clear();
+
+            app.UseIISPlatformHandler();
+
+            app.UseCookieAuthentication(options =>
+            {
+                options.AuthenticationScheme = "Cookies";
+                options.AutomaticAuthenticate = true;
+            });
+
+            app.UseOpenIdConnectAuthentication(options =>
+            {
+                options.AutomaticChallenge = true;
+                options.AuthenticationScheme = "Oidc";
+                options.SignInScheme = "Cookies";
+
+                options.Authority = "http://localhost:5000";
+                options.RequireHttpsMetadata = false;
+
+                options.ClientId = "mvc6";
+                options.ResponseType = "id_token token";
+
+                options.Scope.Add("openid");
+                options.Scope.Add("email");
+                options.Scope.Add("api1");
+            });
+
+            app.UseDeveloperExceptionPage();
+            app.UseMvcWithDefaultRoute();
+        }
+
+        public static void Main(string[] args) => WebApplication.Run<Startup>(args);
+    }
+}
