+++ source/AspNet5/src/MvcClient/Views/Home/CallApi.cshtml
