+++ source/AspNet5/src/MvcClient/Views/Home/Index.cshtml
