--- source/AspNet5Host/NuGet.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <packageSources>
-    <clear />
-    <add key="NuGet" value="https://nuget.org/api/v2/" />
-  </packageSources>
-</configuration>
