--- source/AspNet5Host/global.json
@@ -1,6 +1,0 @@
-﻿{
-  "projects": [ "src", "test" ],
-  "sdk": {
-    "version": "1.0.0-beta8"
-  }
-}
