--- source/AspNet5Host/src/IdentityServerAspNet5/Properties/launchSettings.json
@@ -1,12 +1,0 @@
-{
-  "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
-      "launchBrowser": true,
-      "launchUrl": "https://localhost:44300/",
-      "environmentVariables": {
-        "ASPNET_ENV": "Development"
-      }
-    }
-  }
-}
