--- source/AspNet5Host/src/IdentityServerAspNet5/project.json
@@ -1,35 +1,0 @@
-﻿{
-  "webroot": "wwwroot",
-  "version": "1.0.0-*",
-
-  "dependencies": {
-    "Microsoft.AspNet.Server.Kestrel": "1.0.0-beta8",
-    "Microsoft.AspNet.IISPlatformHandler": "1.0.0-beta8",
-    "Microsoft.AspNet.Diagnostics": "1.0.0-beta8",
-    "Microsoft.Framework.Logging": "1.0.0-beta8",
-    "Microsoft.AspNet.DataProtection": "1.0.0-beta8",
-    "Serilog.Framework.Logging": "1.0.0-beta8-10065",
-    "IdentityServer3.Integration.AspNet": "1.0.0-beta8"
-  },
-
-  "commands": {
-    "web": "Microsoft.AspNet.Server.Kestrel"
-  },
-
-  "frameworks": {
-    "dnx451": { }
-  },
-
-  "publishExclude": [
-    "node_modules",
-    "bower_components",
-    "**.xproj",
-    "**.user",
-    "**.vspscc"
-  ],
-  "exclude": [
-    "wwwroot",
-    "node_modules",
-    "bower_components"
-  ]
-}
